# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Pet API",
    "version": "v1",
    "description": "Unified API for the breeder platform + owner device loop. Routes are being migrated to this spec file-by-file; anything not listed here is still served but not yet documented."
  },
  "servers": [
    {
      "url": "/api/v1"
    },
    {
      "url": "/api",
      "description": "unversioned alias (deprecated)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "RateLimited": {
        "description": "Too many requests. Back off per `Retry-After`.",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds to wait."
          },
          "RateLimit-Limit": {
            "schema": {
              "type": "integer"
            }
          },
          "RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            }
          },
          "RateLimit-Reset": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds until the window resets."
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/breeder/breeding/heat-cycles": {
      "get": {
        "tags": [
          "breeder: breeding"
        ],
        "summary": "Heat cycles (filter by animalId).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "animalId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "heatCycles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    }
                  },
                  "required": [
                    "heatCycles"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      },
      "post": {
        "tags": [
          "breeder: breeding"
        ],
        "summary": "Record a heat cycle.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "animalId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "startedOn": {
                    "type": "string",
                    "minLength": 1
                  },
                  "endedOn": {
                    "nullable": true,
                    "type": "string"
                  },
                  "notes": {
                    "nullable": true,
                    "type": "string"
                  }
                },
                "required": [
                  "animalId",
                  "startedOn"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/breeding/heat-cycles/{id}": {
      "patch": {
        "tags": [
          "breeder: breeding"
        ],
        "summary": "Update a heat cycle (partial).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "404": {
            "description": "not found"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      },
      "delete": {
        "tags": [
          "breeder: breeding"
        ],
        "summary": "Delete a heat cycle.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/breeding/litters/{id}/mated": {
      "post": {
        "tags": [
          "breeder: breeding"
        ],
        "summary": "Record mating detail on a litter (planned → expecting).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "matedOn": {
                    "type": "string",
                    "minLength": 1
                  },
                  "method": {
                    "nullable": true,
                    "type": "string"
                  },
                  "progesterone": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  }
                },
                "required": [
                  "matedOn"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok"
          },
          "404": {
            "description": "litter not found"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/breeding/litters/{id}/progesterone": {
      "post": {
        "tags": [
          "breeder: breeding"
        ],
        "summary": "Append a progesterone reading + get guidance.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "on": {
                    "type": "string",
                    "minLength": 1
                  },
                  "ngml": {
                    "type": "number"
                  }
                },
                "required": [
                  "on",
                  "ngml"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ok"
          },
          "404": {
            "description": "litter not found"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/breeding/calendar": {
      "get": {
        "tags": [
          "breeder: breeding"
        ],
        "summary": "Assembled breeding calendar — heats, litters, go-home dates.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "heats": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "litters": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "goHome": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    }
                  },
                  "required": [
                    "heats",
                    "litters",
                    "goHome"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/breeder/breeding/sweep": {
      "post": {
        "tags": [
          "breeder: breeding"
        ],
        "summary": "Raise season + whelping reminders into the care inbox.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "raised": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "raised"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
...
  }
}